[ixc] pragma parser refactoring
