+++ src/ix.compiler/src/IX.Cs.Compiler/Pragmas/PragmaParser/Ast/AddedPropertyDeclarationAstNode.cs
@@ -1,0 +1,32 @@
+﻿// Ix.Compiler.Cs
+// Copyright (c) 2023 Peter Kurhajec (PTKu), MTS,  and Contributors. All Rights Reserved.
+// Contributors: https://github.com/ix-ax/ix/graphs/contributors
+// See the LICENSE file in the repository root for more information.
+// https://github.com/ix-ax/ix/blob/master/LICENSE
+// Third party licenses: https://github.com/ix-ax/ix/blob/master/notices.md
+
+using Irony.Ast;
+using Irony.Interpreter.Ast;
+using Irony.Parsing;
+
+namespace Ix.Compiler.Cs.Pragmas.PragmaParser;
+
+internal class AddedPropertyDeclarationAstNode : AstNode
+{
+    public string AccessQualifier { get; set; }
+    public string Type { get; set; }
+    public string Identifier { get; set; }
+
+    public override void Init(AstContext context, ParseTreeNode treeNode)
+    {
+        AccessQualifier = treeNode.ChildNodes[2].FindTokenAndGetText();
+        Type = treeNode.ChildNodes[3].FindTokenAndGetText();
+        Identifier = treeNode.ChildNodes[4].FindTokenAndGetText();
+    }
+
+    public override void AcceptVisitor(IAstVisitor visitor)
+    {
+        var v = visitor as PragmaVisitor;
+        v.Product = $"{AccessQualifier} {Type} {Identifier} {{ get; set; }}";
+    }
+}
