+++ src/ix.compiler/src/IX.Cs.Compiler/Pragmas/PragmaParser/Ast/DeclarationAttributeAstNode.cs
@@ -1,0 +1,28 @@
+﻿// Ix.Compiler.Cs
+// Copyright (c) 2023 Peter Kurhajec (PTKu), MTS,  and Contributors. All Rights Reserved.
+// Contributors: https://github.com/ix-ax/ix/graphs/contributors
+// See the LICENSE file in the repository root for more information.
+// https://github.com/ix-ax/ix/blob/master/LICENSE
+// Third party licenses: https://github.com/ix-ax/ix/blob/master/notices.md
+
+using Irony.Ast;
+using Irony.Interpreter.Ast;
+using Irony.Parsing;
+
+namespace Ix.Compiler.Cs.Pragmas.PragmaParser;
+
+internal class DeclarationAttributeAstNode : AstNode
+{
+    public string AttributeLiteral { get; set; }
+
+    public override void Init(AstContext context, ParseTreeNode treeNode)
+    {
+        AttributeLiteral = $"[{treeNode.ChildNodes[2].ChildNodes[1].FindTokenAndGetText()}]";
+    }
+
+    public override void AcceptVisitor(IAstVisitor visitor)
+    {
+        var v = visitor as PragmaVisitor;
+        v.Product = AttributeLiteral;
+    }
+}
